requirements.txt
flask
line-bot-sdk
requests
pymongo
certifi
gunicorn
gradio_client
langchain
transformers
torch
